download {textreadr}	R Documentation

Download Documents

Description

This function enables downloading documents.

Usage

download(url, loc = tempdir(), file.out = NULL)

Arguments

url
The download url(s).

loc
Where to put the files.

file.out
Option vector of names matching url. If this is not given download will try to create a name from url.

Value

Places a copy of the downloaded document in location specified and returns vector of the locations as string paths.

Examples

## Not run: 
m <- download(
c('https://cran.r-project.org/web/packages/curl/curl.pdf',
"https://github.com/trinker/textreadr/raw/master/inst/docs/rl10075oralhistoryst002.pdf"),
)

m

## End(Not run)

[Package textreadr version 0.8.0 Index]