presidential_debates_2012 {textreadr}	R Documentation

2012 U.S. Presidential Debates

Description

A dataset containing a cleaned version of all three presidential debates for the 2012 election.

Usage

data(presidential_debates_2012)

Format

A data frame with 2912 rows and 4 variables

Details

person. The speaker

tot. Turn of talk

dialogue. The words spoken

time. Variable indicating which of the three debates the dialogue is from

[Package textreadr version 0.8.0 Index]